separate test to directories
